+++ src/parser/inline/autolink/hashtag.test.ts
@@ -21,6 +21,11 @@ describe('Unit: parser/inline/autolink/hashtag', () => {
       assert.deepStrictEqual(inspect(parser('##a')), [['##a'], '']);
       assert.deepStrictEqual(inspect(parser('###a')), [['###a'], '']);
       assert.deepStrictEqual(inspect(parser(`#'`)), [[`#'`], '']);
+      assert.deepStrictEqual(inspect(parser(`#a''`)), [[`#a''`], '']);
+      assert.deepStrictEqual(inspect(parser('#_')), [['#_'], '']);
+      assert.deepStrictEqual(inspect(parser('#a_')), [['#a_'], '']);
+      assert.deepStrictEqual(inspect(parser('#a__b')), [['#a__b'], '']);
+      assert.deepStrictEqual(inspect(parser(`#a_'b`)), [[`#a_'b`], '']);
       assert.deepStrictEqual(inspect(parser('#(a)')), [['#'], '(a)']);
       assert.deepStrictEqual(inspect(parser('#{}')), [['#'], '{}']);
       assert.deepStrictEqual(inspect(parser('#{{}')), [['#'], '{{}']);
@@ -49,10 +54,13 @@ describe('Unit: parser/inline/autolink/hashtag', () => {
       assert.deepStrictEqual(inspect(parser('#a(b)')), [['<a class="hashtag" href="/hashtags/a(b)">#a(b)</a>'], '']);
       assert.deepStrictEqual(inspect(parser('#a((b))')), [['<a class="hashtag" href="/hashtags/a">#a</a>'], '((b))']);
       assert.deepStrictEqual(inspect(parser(`#a'`)), [[`<a class="hashtag" href="/hashtags/a'">#a'</a>`], '']);
-      assert.deepStrictEqual(inspect(parser(`#a''`)), [[`<a class="hashtag" href="/hashtags/a">#a</a>`], `''`]);
       assert.deepStrictEqual(inspect(parser(`#a(b')`)), [[`<a class="hashtag" href="/hashtags/a(b')">#a(b')</a>`], '']);
       assert.deepStrictEqual(inspect(parser(`#a(b'')`)), [[`<a class="hashtag" href="/hashtags/a">#a</a>`], `(b'')`]);
       assert.deepStrictEqual(inspect(parser(`#a('b)`)), [['<a class="hashtag" href="/hashtags/a">#a</a>'], `('b)`]);
+      assert.deepStrictEqual(inspect(parser('#a_b')), [['<a class="hashtag" href="/hashtags/a_b">#a_b</a>'], '']);
+      assert.deepStrictEqual(inspect(parser(`#a'_b`)), [[`<a class="hashtag" href="/hashtags/a'_b">#a'_b</a>`], '']);
+      assert.deepStrictEqual(inspect(parser('#a(b_c)')), [['<a class="hashtag" href="/hashtags/a(b_c)">#a(b_c)</a>'], '']);
+      assert.deepStrictEqual(inspect(parser('#a(b__c)')), [['<a class="hashtag" href="/hashtags/a">#a</a>'], '(b__c)']);
       assert.deepStrictEqual(inspect(parser('#あ')), [['<a class="hashtag" href="/hashtags/あ">#あ</a>'], '']);
       assert.deepStrictEqual(inspect(parser('#1a')), [['<a class="hashtag" href="/hashtags/1a">#1a</a>'], '']);
       assert.deepStrictEqual(inspect(parser('#1あ')), [['<a class="hashtag" href="/hashtags/1あ">#1あ</a>'], '']);