+++ src/parser/inline/autolink/hashtag.ts
@@ -1,5 +1,5 @@
 import { AutolinkParser } from '../../inline';
-import { tails, validate, verify, rewrite, creator, open } from '../../../combinator';
+import { tails, verify, rewrite, creator, open } from '../../../combinator';
 import { str } from '../../source';
 import { html } from 'typed-dom';
 
@@ -12,9 +12,8 @@ export const hashtag: AutolinkParser.HashtagParser = creator(rewrite(
       verify(
         str(/^[0-9A-Za-z](?:[0-9A-Za-z-]{0,61}[0-9A-Za-z])?(?:\.[0-9A-Za-z](?:[0-9A-Za-z-]{0,61}[0-9A-Za-z])?)*\//),
         ([source]) => source.length <= 253 + 1),
-      verify(validate(
-        /^[0-9]{0,4}(?:[A-Za-z]|[^\x00-\x7F\s])/,
-        str(/^(?:[0-9A-Za-z]|[^\x00-\x7F\s]|'(?!')){1,128}(?:\((?=[0-9]{0,4}(?:[A-Za-z]|[^\x00-\x7F\s]))(?:[0-9A-Za-z]|[^\x00-\x7F\s]|'(?!')){1,125}\))?/)),
+      verify(
+        str(/^(?=(?:[0-9]{1,4}_?)?(?:[A-Za-z]|[^\x00-\x7F\s]))(?:[0-9A-Za-z]|[^\x00-\x7F\s]|'(?!')|_(?=[0-9A-Za-z]|[^\x00-\x7F\s])){1,128}(?:_?\((?=(?:[0-9]{1,4}_?)?(?:[A-Za-z]|[^\x00-\x7F\s]))(?:[0-9A-Za-z]|[^\x00-\x7F\s]|'(?!')|_(?=[0-9A-Za-z]|[^\x00-\x7F\s])){1,125}\))?(?![0-9A-Za-z'_]|[^\x00-\x7F\s])/),
         ([source]) => source.length <= 128),
     ])),
   (source, { host, url }) => [[